log: fix warning

+++ host-common/log.h
@@ -75,7 +75,7 @@ enum bytes_str_options {
     UPPER_HEX       = 0x20, // Use upper letters for hexadecimal digits
 };
 
-char *bytes_str(const void *in_start, int in_len, const void **in_done, char *out_start, int out_len, int opt);
+char *bytes_str(const void *in_start, size_t in_len, const void **in_done, char *out_start, size_t out_len, int opt);
 
 #define __TRACE(COND, MSG, ...) \
     do {                                                             \
